Add draft submission feature

+++ metadata_backend/api/operators.py
@@ -28,7 +28,7 @@ class BaseOperator(ABC):
                  db_client: AsyncIOMotorClient) -> None:
         """Init needed variables, must be given by subclass.
 
-        :param db_name: Name for database to save files to
+        :param db_name: Name for database to save objects to.
         :param content_type: Content type this operator handles (XML or JSON)
         :param db_client: Motor client used for database connections. Should be
         running on same loop with aiohttp, so needs to be passed from aiohttp
@@ -44,14 +44,34 @@ class BaseOperator(ABC):
         Data formatting and addition step for JSON or XML must be implemented
         by corresponding subclass.
 
-        :param schema_type: Schema type of the object to read.
+        :param schema_type: Schema type of the object to create.
         :param data: Data to be saved to database.
         :returns: Accession id for the object inserted to database
         """
         accession_id = (await self.
                         _format_data_to_create_and_add_to_db(schema_type,
                                                              data))
-        LOG.info(f"Inserting file with schema {schema_type} to database "
+        LOG.info(f"Inserting object with schema {schema_type} to database "
+                 f"succeeded with accession id: {accession_id}")
+        return accession_id
+
+    async def replace_metadata_object(self, schema_type: str,
+                                      accession_id: str,
+                                      data: Union[Dict, str]) -> str:
+        """Replace metadata object from database.
+
+        Data formatting and addition step for JSON or XML must be implemented
+        by corresponding subclass.
+
+        :param schema_type: Schema type of the object to replace.
+        :param accession_id: Identifier of object to replace.
+        :param data: Data to be saved to database.
+        :returns: Accession id for the object inserted to database
+        """
+        await self._format_data_to_replace_and_add_to_db(schema_type,
+                                                         accession_id,
+                                                         data)
+        LOG.info(f"Replacing object with schema {schema_type} to database "
                  f"succeeded with accession id: {accession_id}")
         return accession_id
 
@@ -70,10 +90,11 @@ class BaseOperator(ABC):
         try:
             data_raw = await self.db_service.read(schema_type, accession_id)
             if not data_raw:
+                LOG.error(f"Object with {accession_id} not found.")
                 raise web.HTTPNotFound
             data = await self._format_read_data(schema_type, data_raw)
         except (ConnectionFailure, OperationFailure) as error:
-            reason = f"Error happened while getting file: {error}"
+            reason = f"Error happened while getting object: {error}"
             LOG.error(reason)
             raise web.HTTPBadRequest(reason=reason)
         return data, self.content_type
@@ -83,19 +104,20 @@ class BaseOperator(ABC):
         """Delete metadata object from database.
 
         Tries to remove both JSON and original XML from database, passes
-        silently if files don't exist in database.
+        silently if objects don't exist in database.
 
-        :param schema_type: Schema type of the object to read.
-        :param accession_id: Accession Id of the object to read.
+        :param schema_type: Schema type of the object to delete.
+        :param accession_id: Accession Id of the object to delete.
         :raises: 400 if deleting was not succesful
         """
         db_client = self.db_service.db_client
         await self._remove_object_from_db(Operator(db_client),
                                           schema_type,
                                           accession_id)
-        await self._remove_object_from_db(XMLOperator(db_client),
-                                          schema_type,
-                                          accession_id)
+        if not schema_type.startswith("draft"):
+            await self._remove_object_from_db(XMLOperator(db_client),
+                                              schema_type,
+                                              accession_id)
         LOG.info(f"removing object with schema {schema_type} from database "
                  f"and accession id: {accession_id}")
 
@@ -103,20 +125,54 @@ class BaseOperator(ABC):
                                              data: Dict) -> str:
         """Insert formatted metadata object to database.
 
-        :param schema_type: Schema type of the object to read.
+        :param schema_type: Schema type of the object to insert.
         :param data: Single document formatted as JSON
         :returns: Accession Id for object inserted to database
+        :raises: 400 if reading was not succesful
         """
         try:
             insert_success = (await self.db_service.create(schema_type, data))
         except (ConnectionFailure, OperationFailure) as error:
-            reason = f"Error happened while getting file: {error}"
+            reason = f"Error happened while getting object: {error}"
             LOG.error(reason)
             raise web.HTTPBadRequest(reason=reason)
         if insert_success:
             return data["accessionId"]
         else:
-            reason = "Inserting file to database failed for some reason."
+            reason = "Inserting object to database failed for some reason."
+            LOG.error(reason)
+            raise web.HTTPBadRequest(reason=reason)
+
+    async def _replace_object_from_db(self, schema_type: str,
+                                      accession_id: str,
+                                      data: Dict) -> str:
+        """Replace formatted metadata object in database.
+
+        :param schema_type: Schema type of the object to replace.
+        :param accession_id: Identifier of object to replace.
+        :param data: Single document formatted as JSON
+        :raises: 400 if reading was not succesful, 404 if no data found
+        :returns: Accession Id for object inserted to database
+        """
+        try:
+            check_exists = await self.db_service.exists(schema_type,
+                                                        accession_id)
+            if not check_exists:
+                reason = (f"Object with accession id {accession_id} "
+                          "was not found.")
+                LOG.error(reason)
+                raise web.HTTPNotFound(reason=reason)
+            replace_success = (await self.db_service.replace(schema_type,
+                                                             accession_id,
+                                                             data))
+        except (ConnectionFailure, OperationFailure) as error:
+            reason = f"Error happened while getting object: {error}"
+            LOG.error(reason)
+            raise web.HTTPBadRequest(reason=reason)
+        if replace_success:
+            return accession_id
+        else:
+            reason = "Replacing object to database failed for some reason."
             LOG.error(reason)
             raise web.HTTPBadRequest(reason=reason)
 
@@ -124,11 +180,26 @@ class BaseOperator(ABC):
                                      operator: Any,
                                      schema_type: str,
                                      accession_id: str) -> None:
+        """Delete object from database.
+
+        :param schema_type: Schema type of the object to delete.
+        :param accession_id: Identifier of object to delete.
+        :param data: Single document formatted as JSON
+        :raises: 400 if reading was not succesful, 404 if no data found
+        :returns: None
+        """
         try:
+            check_exists = await operator.db_service.exists(schema_type,
+                                                            accession_id)
+            if not check_exists:
+                reason = (f"Object with accession id {accession_id} "
+                          "was not found.")
+                LOG.error(reason)
+                raise web.HTTPNotFound(reason=reason)
             delete_success = (await operator.db_service.delete(schema_type,
                                                                accession_id))
         except (ConnectionFailure, OperationFailure) as error:
-            reason = f"Error happened while deleting file: {error}"
+            reason = f"Error happened while deleting object: {error}"
             LOG.error(reason)
             raise web.HTTPBadRequest(reason=reason)
         if delete_success:
@@ -142,6 +213,15 @@ class BaseOperator(ABC):
     async def _format_data_to_create_and_add_to_db(self, schema_type: str,
                                                    data: Any) -> str:
         """Format and add data to database.
+
+        Must be implemented by subclass.
+        """
+
+    @abstractmethod
+    async def _format_data_to_replace_and_add_to_db(self, schema_type: str,
+                                                    accession_id: str,
+                                                    data: Any) -> str:
+        """Format and replace data in database.
 
         Must be implemented by subclass.
         """
@@ -183,7 +263,7 @@ class Operator(BaseOperator):
         :param page_size: Results per page
         :param page_num: Page number
         :raises: HTTPBadRequest if error happened when connection to database
-        and HTTPNotFound error if file with given accession id is not found.
+        and HTTPNotFound error if object with given accession id is not found.
         :returns: Query result with pagination numbers
         """
         # Generate mongodb query from query parameters
@@ -211,21 +291,21 @@ class Operator(BaseOperator):
         try:
             cursor = self.db_service.query(schema_type, mongo_query)
         except (ConnectionFailure, OperationFailure) as error:
-            reason = f"Error happened while getting file: {error}"
+            reason = f"Error happened while getting object: {error}"
             LOG.error(reason)
             raise web.HTTPBadRequest(reason=reason)
         skips = page_size * (page_num - 1)
         cursor.skip(skips).limit(page_size)
         data = await self._format_read_data(schema_type, cursor)
         if not data:
-            LOG.error("could not find any data.")
+            LOG.error(f"could not find any data in {schema_type}.")
             raise web.HTTPNotFound
         page_size = len(data) if len(data) != page_size else page_size
         total_objects = await self.db_service.get_count(schema_type,
                                                         mongo_query)
         LOG.debug(f"DB query: {que}")
-        LOG.info(f"DB query successful for query on {schema_type}"
-                 f"resulted in {total_objects}."
+        LOG.info(f"DB query successful for query on {schema_type} "
+                 f"resulted in {total_objects}. "
                  f"Requested was page {page_num} and page size {page_size}.")
         return data, page_num, page_size, total_objects
 
@@ -239,7 +319,7 @@ class Operator(BaseOperator):
         By default date is two months from submission date (based on ENA
         submission model).
 
-        :param schema_type: Schema type of the object to read.
+        :param schema_type: Schema type of the object to create.
         :param data: Metadata object
         :returns: Accession Id for object inserted to database
         """
@@ -249,8 +329,37 @@ class Operator(BaseOperator):
         data["dateModified"] = datetime.utcnow()
         if schema_type == "study":
             data["publishDate"] = datetime.utcnow() + relativedelta(months=2)
-        LOG.debug(f"Operator formatted data for {schema_type} to add to DB")
+        LOG.debug(f"Operator formatted data for {schema_type} to add to DB.")
         return await self._insert_formatted_object_to_db(schema_type, data)
+
+    async def _format_data_to_replace_and_add_to_db(self, schema_type: str,
+                                                    accession_id: str,
+                                                    data: Dict) -> str:
+        """Format JSON metadata object and replace it in db.
+
+        Replace information to object before adding to db.
+
+        We will not replace accessionId, publishDate or dateCreated,
+        as these should are generated when created.
+
+        We will keep also publisDate and dateCreated from old object.
+
+        :param schema_type: Schema type of the object to replace.
+        :param accession_id: Identifier of object to replace.
+        :param data: Metadata object
+        :returns: Accession Id for object inserted to database
+        """
+        forbidden_keys = ['accessionId', 'publishDate', 'dateCreated']
+        if any([i in data for i in forbidden_keys]):
+            reason = (f"Some items (e.g: {', '.join(forbidden_keys)})"
+                      "cannot be changed.")
+            LOG.error(reason)
+            raise web.HTTPBadRequest(reason=reason)
+        data["accessionId"] = accession_id
+        data["dateModified"] = datetime.utcnow()
+        LOG.debug(f"Operator formatted data for {schema_type} to add to DB")
+        return await self._replace_object_from_db(schema_type, accession_id,
+                                                  data)
 
     def _generate_accession_id(self) -> str:
         """Generate random accession id.
@@ -341,6 +450,21 @@ class XMLOperator(BaseOperator):
                 _insert_formatted_object_to_db(schema_type,
                                                {"accessionId": accession_id,
                                                 "content": data}))
+
+    async def _format_data_to_replace_and_add_to_db(self, schema_type: str,
+                                                    accession_id: str,
+                                                    data: str) -> str:
+        """Format XML metadata object and add it to db.
+
+        XML is validated, then parsed to json and json is added to database.
+        After successful json insertion, xml itself is backed up to database.
+
+        :param schema_type: Schema type of the object to replace.
+        :param accession_id: Identifier of object to replace.
+        :param data: Original xml content
+        :returns: Accession Id for object inserted to database
+        """
+        raise web.HTTPNotImplemented
 
     async def _format_read_data(self, schema_type: str, data_raw: Dict) -> str:
         """Get XML content from given mongodb data.
